Update precision and scale validation to decimal variables

+++ db/migrate/20180808160105_create_stock_ownerships.rb
@@ -1,10 +1,10 @@
 class CreateStockOwnerships < ActiveRecord::Migration[5.2]
   def change
     create_table :stock_ownerships do |t|
-      t.float :price_bought, null: false, scale: 2
+      t.decimal :price_bought, null: false, precision: 7, scale: 2
       t.integer :amount, null: false
-      t.float :high_range, scale: 2
-      t.float :low_range, scale: 2
+      t.decimal :high_range, precision: 7, scale: 2
+      t.decimal :low_range, precision: 7, scale: 2
 
       t.belongs_to :user, null: false
       t.belongs_to :stock, null: false
